Revert "AC-392 - Download patient option not switching instantly (#7)"
This reverts commit 7393e0b7de3962e8f734123cf869a95858b3a644.
